% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{interleave}
\usage{
interleave(x, y, z = NULL)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector. To be interleaved with x.}

\item{z}{If provided, third vector. To be interleaved with x and y.}
}
\value{
A vector of length length(x) + length(y) with the elements of x and y interleaved.
}
\description{
Interleave two (or three) vectors.
}
