% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gdsc.R
\name{merge_gdsc}
\alias{merge_gdsc}
\title{merge_gdsc}
\usage{
merge_gdsc(
  data,
  drugsA = c("1014", "1036", "1058", "156"),
  drugsB = c("1526", "1371", "1527", "1066")
)
}
\arguments{
\item{data}{Must have COSMIC_ID and DRUG_ID columns. Likely the output of reps_gdsc.}

\item{drugsA}{Vector of DRUG_IDs for replicate A.}

\item{drugsB}{Vector of DRUG_IDs for replicate B.}
}
\value{
A data frame containing one row for each drug-CL combination for replicates A and B.
}
\description{
Creates a data frame with a single row for each replicated drug-CL combination
}
