% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(plotlist = NULL, cols = 1, title = NULL)
}
\arguments{
\item{plotlist}{List of plots.}

\item{cols}{Number of columns in the grid of plots to produce}

\item{title}{Main title.}
}
\value{
A grid of plots.
}
\description{
Print out several plots in a grid.
}
