% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDR_ccle.R
\name{plotDR_ccle}
\alias{plotDR_ccle}
\title{plotDR_ccle}
\usage{
plotDR_ccle(
  data = norm_ccle,
  combo1,
  combo2,
  manScale = NA,
  title = "CCLE",
  connect = FALSE
)
}
\arguments{
\item{data}{Data frame to plot. The x-axis will display log2(CONCENTRATION); the y-axis will display RV_MUC.}

\item{combo1}{The COMBO_ID of the drug-CL combination to plot.}

\item{combo2}{Optional. A second COMBO_ID so that a replicate of the drug-CL combination can added to the plot.}

\item{manScale}{An optional upper y-axis limit.}

\item{title}{Plot title.}

\item{connect}{Logical. Should the plotted points be connected by a line?}
}
\value{
A CCLE dose-response plot.
}
\description{
Plot a dose-response curve for one (or two) CCLE drug-cell line combinations.
}
