% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDR_gdsc.R
\name{plotDR_gdsc}
\alias{plotDR_gdsc}
\title{plotDR_gdsc}
\usage{
plotDR_gdsc(
  data = norm_gdsc,
  CL,
  drug1,
  drug2 = NA,
  manScale = NA,
  title = "GDSC",
  connect = FALSE
)
}
\arguments{
\item{data}{Data frame to plot. The x-axis will display log2(CONC); the y-axis will display RV_MUC.}

\item{CL}{The COSMIC_ID of the drug_CL combination to plot.}

\item{drug1}{The DRUG_ID of the drug_CL combination to plot.}

\item{drug2}{Optional. A second DRUG_ID so that a replicate of the drug_CL combination can added to the plot.}

\item{manScale}{An optional upper y-axis limit.}

\item{title}{Plot title.}

\item{connect}{Logical. Should the plotted points be connected by a line?}
}
\value{
A GDSC dose-response plot.
}
\description{
Plot a dose-response curve for one (or two) GDSC drug-cell line combinations.
}
