% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateFormat_ccle.R
\name{plotFormat_ccle}
\alias{plotFormat_ccle}
\title{plotFormat_ccle}
\usage{
plotFormat_ccle(data = raw_ccle, plate)
}
\arguments{
\item{data}{Plate format data frame.}

\item{plate}{The ASSAY_PLATE_NAME of the plate layout to plot.}
}
\value{
A CCLE plate format plot.
}
\description{
Display the plate format for a CCLE plate layout.
}
