% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateFormat_gdsc.R
\name{plotFormat_gdsc}
\alias{plotFormat_gdsc}
\title{plotFormat_gdsc}
\usage{
plotFormat_gdsc(data = gdsc_plate_maps, drugset)
}
\arguments{
\item{data}{Plate format data frame.}

\item{drugest}{The DRUGSET_ID of the plate layout to plot.}
}
\value{
A GDSC plate format plot.
}
\description{
Display the plate format for a GDSC plate layout.
}
