% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeat_ccle.R
\name{plotHeat_ccle}
\alias{plotHeat_ccle}
\title{plotHeat_ccle}
\usage{
plotHeat_ccle(data = raw_ccle, scan, legendLoc = "side")
}
\arguments{
\item{data}{Data frame to plot.}

\item{scan}{The ASSAY_PLATE_NAME of the plate to plot.}

\item{legendLoc}{The location of the legend, either "side" or "bottom".}
}
\value{
A CCLE heatmap.
}
\description{
Plot a heatmap for one CCLE plate.
}
