% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeat_gdsc.R
\name{plotHeat_gdsc}
\alias{plotHeat_gdsc}
\title{plotHeat_gdsc}
\usage{
plotHeat_gdsc(
  data = raw_gdsc,
  scan = NA,
  drug = NA,
  CL = NA,
  legendLoc = "side"
)
}
\arguments{
\item{data}{Data frame to plot.}

\item{scan}{The SCAN_ID of the plate to plot. Either scan or drug and CL must be used.}

\item{drug}{May specify the plate by providing both the DRUG_ID and COSMIC_ID of a drug-CL combination.}

\item{CL}{May specify the plate by providing both the DRUG_ID and COSMIC_ID of a drug-CL combination.}

\item{legendLoc}{The location of the legend, either "side" or "bottom".}
}
\value{
A GDSC heatmap.
}
\description{
Plot a heatmap for one GDSC plate.
}
