% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps_gdsc.R
\name{reps_gdsc}
\alias{reps_gdsc}
\title{reps_gdsc}
\usage{
reps_gdsc(
  data,
  d = 0.5,
  data2 = gdsc_nlme_stats,
  drugsA = c("1014", "1036", "1058", "156"),
  drugsB = c("1526", "1371", "1527", "1066")
)
}
\arguments{
\item{data}{Must have COSMIC_ID and DRUG_ID columns and no maxc column.}

\item{d}{Allowed difference between replicates' max concentration.}

\item{data2}{Must contain max concentration.}

\item{drugsA}{Vector of DRUG_IDs for replicate A.}

\item{drugsB}{Vector of DRUG_IDs for replicate B.}
}
\value{
A data frame containing only observations for the given replicates that were tested at the same concentrations in both replicate A and B.
}
\description{
Select rows for replicates tested at the same maxc.
}
